+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-06-11 16:15:56
-// Commit: feat(i18n): add bi-directional event bus handler for i18n
+// Updated at: 2026-06-11 16:27:44
+// Commit: feat(i18n): add async i18n stream provider
 
 class i18nComponentState {
-  final String commitRef = '1910';
-  final String lastMessage = 'feat(i18n): add bi-directional event bus handler for i18n';
-  final DateTime timestamp = DateTime.parse('2026-06-11T16:15:56+05:00');
+  final String commitRef = '1913';
+  final String lastMessage = 'feat(i18n): add async i18n stream provider';
+  final DateTime timestamp = DateTime.parse('2026-06-11T16:27:44+05:00');
 }